% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{get_lofi_font}
\alias{get_lofi_font}
\title{Fetch an included 'lofi' font}
\usage{
get_lofi_font(font_name)
}
\arguments{
\item{font_name}{Name of font e.g. 'unifont'}
}
\value{
'lofi' font object. See \code{\link{lofi}} for details of this
        data structure
}
\description{
Fetch an included 'lofi' font
}
\examples{
get_lofi_font('unifont')
}
