% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{get_lofi_names}
\alias{get_lofi_names}
\title{Return the names of all included fonts}
\usage{
get_lofi_names(type)
}
\arguments{
\item{type}{font type. Either 'bitmap', 'vector' or 'all'}
}
\value{
List of two elements: names of bitmap fonts, names of vector fonts
}
\description{
Return the names of all included fonts
}
\examples{
get_lofi_names('bitmap')
}
