% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-info.R
\name{get_lofifont_names}
\alias{get_lofifont_names}
\title{#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#' List of included bitmap fonts. 
#' 
#' Each font is an object with class 'lofifont'
#' 
#' \describe{
#'   \item{coords}{data.frame}
#' }
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
"bitmap_fonts"}
\usage{
get_lofifont_names()
}
\value{
List of two elements: names of bitmap fonts, names of vector fonts
}
\description{
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#' List of included vector fonts. 
#' 
#' Each font is an object with class 'lofifont'
#' 
#' \describe{
#'   \item{coords}{data.frame}
#' }
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
"vector_fonts"
Return the names of all included fonts
}
\examples{
get_lofifont_names()
}
