% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofifonts-package.R
\docType{package}
\name{lofifonts-package}
\alias{lofifonts}
\alias{lofifonts-package}
\title{lofifonts: Text Rendering with Bitmap and Vector Fonts}
\description{
Alternate font rendering is useful when rendering to novel graphics outputs where modern font rendering is not available or not appropriate. Bitmap and vector fonts allow for bespoke rendering using pixel coordinates and line drawing. A selection of fonts is included and all can be rendered to a matrix of pixel locations or returned as pixel coordinates and collections of stroke endpoints.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/lofifonts}
  \item Report bugs at \url{https://github.com/coolbutuseless/lofifonts/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

Authors:
\itemize{
  \item Frederic Cambus \email{fred@statdns.com} (Creator of 'spleen' font) [copyright holder]
  \item Anders Hoff (Creator of 'gridfont' font) [copyright holder]
}

Other contributors:
\itemize{
  \item GNU Unifont authors (Creators of 'unifont') [copyright holder]
}

}
\keyword{internal}
