% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.lofi-raster}
\alias{plot.lofi-raster}
\title{Plot a lofi raster}
\usage{
\method{plot}{`lofi-raster`}(x, interpolate = FALSE, ...)
}
\arguments{
\item{x}{lofi raster rendering}

\item{interpolate}{default: FALSE}

\item{...}{extra arguments passed to \code{plot()}}
}
\value{
None
}
\description{
Plot a lofi raster
}
\examples{
ras <- bitmap_text_raster("Hi")
plot(ras)
}
