% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lofi}
\alias{print.lofi}
\title{Print summary information about a lofi font}
\usage{
\method{print}{lofi}(x, ...)
}
\arguments{
\item{x}{lofi font object}

\item{...}{other arguments ignored}
}
\value{
None
}
\description{
Print summary information about a lofi font
}
\examples{
font <- get_lofi_font('unscii-8')
print(font)
}
