% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lofifont}
\alias{print.lofifont}
\title{Print summary information about a lofi font}
\usage{
\method{print}{lofifont}(x, ...)
}
\arguments{
\item{x}{lofi font object}

\item{...}{other arguments ignored}
}
\description{
Print summary information about a lofi font
}
