% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\name{bitmap_text_coords}
\alias{bitmap_text_coords}
\title{Create a data.frame of pixel coordinate information of the rendered text}
\usage{
bitmap_text_coords(text, font = "unifont", line_height = NULL, missing = NULL)
}
\arguments{
\item{text}{Single text string. Can include carriage returns.}

\item{font}{Name of bitmap font. One of the following:
\itemize{
  \item{spleen: "spleen-12x24", "spleen-16x32", "spleen-32x64", "spleen-5x8", "spleen-6x12", "spleen-8x16"}
  \item{tamzen: "Tamzen10x20b", "Tamzen10x20r", "Tamzen5x9b", "Tamzen5x9r", "Tamzen6x12b", "Tamzen6x12r", "Tamzen7x13b", "Tamzen7x13r", "Tamzen7x14b", "Tamzen7x14r", "Tamzen8x15b", "Tamzen8x15r", "Tamzen8x16b", "Tamzen8x16r"}
  \item{"unifomt" (the default)}
  \item{unscii: "unscii-8", "unscii-8-thin"}
}}

\item{line_height}{Integer value for the vertical distance between multiple lines
of text.  Use this to override the font's lineheight.
Default: NULL means to use the font's built-in lineheight.}

\item{missing}{Codepoint (integer) to use if glyph not found in font. 
Default: NULL means to use the default specified by the font internally.
Otherwise it will default to the codepoint for '?'}
}
\value{
data.frame of coordinate information
\describe{
  \item{\code{char_idx}}{The index of the character within the provided \code{text} string}
  \item{\code{codepoint}}{Unicode codepoint (integer)}
  \item{\code{x}}{Pixel coordinate x value for display}
  \item{\code{y}}{Pixel coordinate y value for display}
  \item{\code{line}}{Line number within input \code{text} where this character appears}
  \item{\code{x0}}{Original untransformed x-coordinate}
  \item{\code{y0}}{Original untransformed y-coordinate}
}
}
\description{
Create a data.frame of pixel coordinate information of the rendered text
}
\examples{
bitmap_text_coords('Hi')
}
\seealso{
Other bitmap text functions: 
\code{\link{bitmap_text_matrix}()},
\code{\link{bitmap_text_raster}()}
}
\concept{bitmap text functions}
