% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\name{bitmap_text_raster}
\alias{bitmap_text_raster}
\title{Create a raster image of the rendered text}
\usage{
bitmap_text_raster(
  text,
  font = "unifont",
  dx = 0L,
  dy = 0L,
  scale = 1,
  missing = NULL
)
}
\arguments{
\item{text}{Single text string. Can include carriage returns.}

\item{font}{Name of bitmap font. One of the following:
\itemize{
  \item{spleen: "spleen-12x24", "spleen-16x32", "spleen-32x64", "spleen-5x8", "spleen-6x12", "spleen-8x16"}
  \item{tamzen: "Tamzen10x20b", "Tamzen10x20r", "Tamzen5x9b", "Tamzen5x9r", "Tamzen6x12b", "Tamzen6x12r", "Tamzen7x13b", "Tamzen7x13r", "Tamzen7x14b", "Tamzen7x14r", "Tamzen8x15b", "Tamzen8x15r", "Tamzen8x16b", "Tamzen8x16r"}
  \item{"unifomt" (the default)}
  \item{unscii: "unscii-8", "unscii-8-thin"}
}}

\item{dx}{Additional character spacing in the horizontal direction. Default: 0}

\item{dy}{Additional character spacing in the vertical direction. Default: 0}

\item{scale}{Integer size scale factor. Default: 1.  Must be an integer value >= 1.
Scale up the matrix or raster result by this factor}

\item{missing}{Codepoint (integer) to use if glyph not found in font. 
Default: NULL means to use the default specified by the font internally.
Otherwise it will default to the codepoint for '?'}
}
\value{
Raster image representation of the rendered text
}
\description{
Create a raster image of the rendered text
}
\examples{
ras <- bitmap_text_raster('Hi')
plot(ras, interpolate = FALSE)
}
\seealso{
Other bitmap text functions: 
\code{\link{bitmap_text_coords}()},
\code{\link{bitmap_text_matrix}()}
}
\concept{bitmap text functions}
