% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofifonts-package.R
\docType{package}
\name{lofifonts-package}
\alias{lofifonts}
\alias{lofifonts-package}
\title{lofifonts: Text Rendering with Bitmap and Vector Fonts}
\description{
Alternate font rendering is useful when rendering text to novel graphics outputs where modern font rendering is not available or where bespoke text positioning is required. Bitmap and vector fonts allow for custom layout and rendering using pixel coordinates and line drawing. Formatted text is created as a data.frame of pixel coordinates (for bitmap fonts) or stroke coordinates (for vector fonts). All text can be easily previewed as a matrix or raster image. A selection of fonts is included with this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/lofifonts}
  \item Report bugs at \url{https://github.com/coolbutuseless/lofifonts/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

Other contributors:
\itemize{
  \item June Choe \email{jchoe001@gmail.com} (Contributed character positioning code) [contributor]
  \item Frederic Cambus \email{fred@statdns.com} (Creator of 'spleen' font) [copyright holder, type designer]
  \item GNU Unifont authors (Creators of 'unifont') [copyright holder, type designer]
  \item Suraj Kurapati (Creator of 'Tamzen' font) [copyright holder, type designer]
  \item Scott Fial (Creator of 'Tamsyn' font upon which 'Tamzen' is based) [copyright holder, type designer]
  \item Anders Hoff (Creator of 'gridfont' font) [copyright holder, type designer]
}

}
\keyword{internal}
