% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-text.R
\name{vector_text_coords}
\alias{vector_text_coords}
\title{Create data.frame of glyph information for the given text.}
\usage{
vector_text_coords(
  text,
  font = c("gridfont", "gridfont_smooth", "arcade"),
  dx = 0,
  dy = 0,
  missing = utf8ToInt("?"),
  line_height = NULL
)
}
\arguments{
\item{text}{Single text string. Can include carriage returns.}

\item{font}{Name of vector font. One of 
\code{c("arcade", "gridfont", "gridfont_smooth")}}

\item{dx}{Additional character spacing in the horizontal direction. Default: 0}

\item{dy}{Additional character spacing in the vertical direction. Default: 0}

\item{missing}{Codepoint to use if glyph not available in font. default: Codepoint
for '?'}

\item{line_height}{line height}
}
\value{
data.frame of stroke information
\describe{
  \item{\code{char_idx}}{The index of the character within the provided \code{text} string}
  \item{\code{codepoint}}{Unicode codepoint (integer)}
  \item{\code{stroke_idx}}{Index of the stroke within each character}
  \item{\code{point_idx}}{Index of the point within each stroke}
  \item{\code{x}}{Pixel coordinate x value for display}
  \item{\code{y}}{Pixel coordinate y value for display}
  \item{\code{width}}{Width of vector character}
  \item{\code{height}}{Height of vector character}
  \item{\code{x0}}{Original untransformed x-coordinate}
  \item{\code{y0}}{Original untransformed y-coordinate}
  \item{\code{line}}{Line number within input \code{text} where this character appears}
}
}
\description{
Text input can contain multiple lines separated by carriage returns
}
\examples{
vector_text_coords('Hi')
}
\seealso{
Other vector text functions: 
\code{\link{vector_text_matrix}()},
\code{\link{vector_text_raster}()}
}
\concept{vector text functions}
