% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-text.R
\name{vector_text_matrix}
\alias{vector_text_matrix}
\title{Create a binary matrix of the rendered text}
\usage{
vector_text_matrix(
  text,
  font = c("gridfont", "gridfont_smooth", "arcade"),
  scale = 1,
  dx = NULL,
  dy = NULL,
  missing = utf8ToInt("?")
)
}
\arguments{
\item{text}{Single text string. Can include carriage returns.}

\item{font}{Name of vector font. One of 
\code{c("arcade", "gridfont", "gridfont_smooth")}}

\item{scale}{Scale factor for text rendering. Numeric value greater than zero.
Default: 1}

\item{dx}{Additional character spacing in the horizontal direction. Default: 0}

\item{dy}{Additional character spacing in the vertical direction. Default: 0}

\item{missing}{Codepoint to use if glyph not available in font. default: Codepoint
for '?'}
}
\value{
Binary matrix rendering of the font
}
\description{
Create a binary matrix of the rendered text
}
\examples{
vector_text_matrix("Hi")
}
\seealso{
Other vector text functions: 
\code{\link{vector_text_coords}()},
\code{\link{vector_text_raster}()}
}
\concept{vector text functions}
